adding columns to display
added columns for the hour slots to be displayed on column 1 of the page.

added on column 10 for the save button to be displayed and appended the class details to the for loop to be appended to the container to display. appended all classes together for text and buttons to display appropriately in correct location. grabbed "far fa-save" from documentation in the font awesome licensing page as a reference.