variable and function for color
Added a global variable for the function of background coloring depending on what time it currently is.

functioned using moment and hours from HourSign for each time block  and parsed it into the id attribute

added if, else if and else to denote which color should be applied to which time block depending on what hour it currently is.