adding local storage functionality/ fixing background color
fixed issue with hourSign being in the function for color time that prevented the correct background color to be displayed. gave each else and else if separate lines to read code easier.

Added value to text areas as well as a function to get the information that is stored.

added button function for on click saving purposes when clicking save.

added set items coordinating all the text areas and values set from 0-8 to be displayed in the appropriate block on the page after reloading.